+++ app/src/main/java/com/navapp/navigation/util/exceptions/LackingPermissionException.java
@@ -1,4 +1,4 @@
-package com.navapp.navigation.util.geofence;
+package com.navapp.navigation.util.exceptions;
 
 public class LackingPermissionException extends Exception {
     public LackingPermissionException(String permissionName) {
